Fix LoadError during startup with Redmine trunk r21288 #33
