+++ app/models/filter_rule.rb
@@ -20,7 +20,7 @@ class FilterRule < Setting
 
     remote_ip_addr = IPAddr.new(remote_ip)
 
-    always_allowed_ip_list = IPFilterConfig['always_allowed_ip_list'] || []
+    always_allowed_ip_list = RedmineIpFilter::IpFilterConfig['always_allowed_ip_list'] || []
     (self.allowed_ip_list | always_allowed_ip_list).any? do |ip|
       begin
         IPAddr.new(ip).include?(remote_ip_addr)
